refactor: Improve Stripe session handling and return route
- Rename stripeSession.tsx to session.tsx for clarity
- Add server-side URL generation for return URL
- Update CheckoutForm import to use new session file
- Create dedicated return route for handling Stripe checkout sessions
